backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.7
docker==7.1.0
python-dotenv==1.0.1
